requirements.txt
pathway>=0.4.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
llama-index>=0.8.0.post1
llama-index>=0.8.0
pandas>=1.5.0
scikit-learn>=1.2.0s